pyproject.toml
[project]
name = "cook-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
]
